+++ src/OpenTelemetry/CHANGELOG.md
@@ -32,12 +32,12 @@ Released 2023-Jul-12
     API](https://github.com/open-telemetry/opentelemetry-specification/blob/976432b74c565e8a84af3570e9b82cb95e1d844c/specification/logs/bridge-api.md)
     implementation (`Sdk.CreateLoggerProviderBuilder`, etc.).
     ([#4433](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4433))
-  
+
   * Obsoleted `LogRecord.LogLevel` in favor of the `LogRecord.Severity` property
     which matches the [OpenTelemetry Specification > Logs DataModel > Severity
     definition](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/logs/data-model.md#field-severitynumber).
     ([#4433](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4433))
-  
+
   * Added `LogRecord.Logger` property to access the [OpenTelemetry Specification
     Instrumentation
     Scope](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/glossary.md#instrumentation-scope)